+++ filepanther/filepath_to_metadata_test.py
@@ -73,13 +73,14 @@ class Test_filepath_to_metadata(TestCase):
     #     self.assertEqual(result, {
     #         "this": "abc"
     #     })
-    #
-    # def test_duplicate_strftime_directive(self):
-    #     """can parse string with duplicate strftime directive"""
-    #     result = filepath_to_metadata(
-    #         format_string="%Y_duplicate_thing_%Y.txt",
-    #         filepath="2020_one_thing_2020.txt"
-    #     )
-    #     self.assertEqual(result, {
-    #         "this": "abc"
-    #     })
+
+    def test_duplicate_strftime_directive(self):
+        """can parse string with duplicate strftime directive"""
+        result = filepath_to_metadata(
+            format_string="%Y_duplicate_thing_%Y.txt",
+            filepath="2020_duplicate_thing_2020.txt"
+        )
+        self.assertEqual(result, {
+            "dt_Y": 2020,
+            "_datetime": datetime(2020, 1, 1)
+        })